+++ lib/test/integration/kafka_test.go
@@ -78,6 +78,8 @@ output:
     topic: topic-$ID
     max_in_flight: $MAX_IN_FLIGHT
     retry_as_batch: $VAR3
+    metadata:
+      exclude_prefixes: [ $OUTPUT_META_EXCLUDE_PREFIX ]
     batching:
       count: $OUTPUT_BATCH_COUNT
 
@@ -95,6 +97,7 @@ input:
 	suite := integrationTests(
 		integrationTestOpenClose(),
 		integrationTestMetadata(),
+		integrationTestMetadataFilter(),
 		integrationTestSendBatch(10),
 		integrationTestStreamSequential(1000),
 		integrationTestStreamParallel(1000),